fix: change int datat to decimal for many value

+++ src/models/jobs.py
@@ -12,8 +12,8 @@ class Job(Base):
     user_id = sa.Column(sa.Integer, sa.ForeignKey('users.id'), comment="Идентификатор пользователя")
     title = sa.Column(sa.String, comment="Название вакансии")
     description = sa.Column(sa.String, comment="Описание ввкансии")
-    salary_from = sa.Column(sa.Integer, comment="Зарплата от")
-    salary_to = sa.Column(sa.Integer, comment="Зарплата до")
+    salary_from = sa.Column(sa.DECIMAL, comment="Зарплата от")
+    salary_to = sa.Column(sa.DECIMAL, comment="Зарплата до")
     is_active = sa.Column(sa.Boolean, default=False)
     created_at = sa.Column(sa.DateTime, default=datetime.datetime.utcnow)
